@@ -1,5 +1,6 @@
-import { Component, OnInit } from '@angular/core';
-import { Router } from "@angular/router";
+import { Component } from '@angular/core';
+import { Router } from '@angular/router';
+import { faBars, faXmark } from '@fortawesome/free-solid-svg-icons';
 import { LoginService } from 'src/app/services/login.service';
 
 @Component({
@@ -7,17 +8,22 @@ import { LoginService } from 'src/app/services/login.service';
   templateUrl: './header.component.html',
   styleUrls: ['./header.component.css']
 })
-export class HeaderComponent implements OnInit {
+export class HeaderComponent {
+  menuIcon = faBars
+  cancleIcon = faXmark
+  menuOption = faBars
+
+  open(menu: HTMLDivElement) {
+    menu.classList.toggle("-translate-y-[50vh]")
+    this.menuOption = this.menuOption == this.menuIcon ? this.cancleIcon : this.menuIcon
+  }
 
   constructor(private router: Router, private lin: LoginService) {
     lin.auth().subscribe((data: any) => {
       if (data['valid'] === 'true') {
-        this.router.navigate(['/', 'users'])
+        this.router.navigate(['/', 'chat'])
       }
     })
-  }
-
-  ngOnInit(): void {
   }
 
 }
